update move actions
